+++ vclib-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oidc/AuthenticationRequestParameters.kt
@@ -235,6 +235,19 @@ data class AuthenticationRequestParameters(
      */
     @SerialName("resource")
     val resource: String? = null,
+
+    /**
+     * RFC7636: A challenge derived from the code verifier that is sent in the authorization request, to be verified
+     * against later.
+     */
+    @SerialName("code_challenge")
+    val codeChallenge: String? = null,
+
+    /**
+     * RFC7636: A method that was used to derive code challenge.
+     */
+    @SerialName("code_challenge_method")
+    val codeChallengeMethod: String? = null,
 ) {
 
     fun serialize() = jsonSerializer.encodeToString(this)