+++ vclib-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/SerializationTest.kt
@@ -7,7 +7,7 @@ import at.asitplus.wallet.lib.oidc.OpenIdConstants.TOKEN_TYPE_BEARER
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldContain
-import io.ktor.http.encodeURLParameter
+import io.ktor.http.*
 import kotlinx.serialization.encodeToString
 import kotlin.random.Random
 
@@ -53,7 +53,9 @@ class SerializationTest : FunSpec({
 
     fun createCredentialRequest() = CredentialRequestParameters(
         format = CredentialFormatEnum.JWT_VC,
-        types = arrayOf(randomString(), randomString()),
+        credentialDefinition = SupportedCredentialFormatDefinition(
+            types = listOf(randomString(), randomString()),
+        ),
         proof = CredentialRequestProof(
             proofType = randomString(),
             jwt = randomString()
@@ -125,8 +127,8 @@ class SerializationTest : FunSpec({
         val params = createCredentialRequest()
         val json = at.asitplus.wallet.lib.oidc.jsonSerializer.encodeToString(params)
         println(json)
-        json shouldContain "\"types\":["
-        json shouldContain "\"${params.types.first()}\""
+        json shouldContain "\"type\":["
+        json shouldContain "\"${params.credentialDefinition?.types?.first()}\""
         val parsed: CredentialRequestParameters =
             at.asitplus.wallet.lib.oidc.jsonSerializer.decodeFromString<CredentialRequestParameters>(json)
         parsed shouldBe params