+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Release 3.6.0:
      - Ktor 2.3.10
      - Auto-publish version catalogs
  - Implement OpenID for Verifiable Credential Issuance draft 13, from 2024-02-08
+ - Implement RFC 7636 Proof Key for Code Exchange for OpenID for Verifiable Credential Issuance implementations, i.e. `IssuerService` and `WalletService`
+ - `IssuerService`: Make public API functions suspending
+ - `WalletService`: Make public API functions suspending
  - TODO document changes in `IssuerService` and `WalletService`
 
 Release 3.5.0: